Saving and loading of training completion
