+++ packages/frontend/src/components/PracticeActivity.component.tsx
@@ -1,35 +1,12 @@
 import { ActivityCard } from './ActivityCard.component';
-import { FC, useEffect, useState } from 'react';
+import { FC } from 'react';
 import cameraCheckImage from '../assets/TaskPreviewImage.png';
 import { useNavigate } from 'react-router';
-import { usersControllerIsTrainingComplete } from '../client';
-import { useUser } from '../contexts/User.context';
+import { useTraining } from '../contexts/Training.contenxt';
 
 export const PracticeActivity: FC = () => {
   const navigate = useNavigate();
-  const [hasCompleted, setHasCompleted] = useState<boolean>(false);
-  const { user } = useUser();
-
-  const getHasCompletedTraining = async () => {
-    const hasCompletedResults = await usersControllerIsTrainingComplete({
-      path: {
-        id: user!.id
-      }
-    });
-
-    if (hasCompletedResults.error || !hasCompletedResults.data) {
-      console.error(hasCompletedResults.error);
-      return;
-    }
-
-    setHasCompleted(hasCompletedResults.data.complete);
-  };
-
-
-  useEffect(() => {
-    getHasCompletedTraining();
-  }, []);
-
+  const { hasCompletedTraining } = useTraining();
 
   return (
     <ActivityCard
@@ -37,7 +14,7 @@ export const PracticeActivity: FC = () => {
       activityTitle="Practice Task: Model Fractions"
       activityDescription='Click "start" to begin a brief, required tutorial about this site.'
       activityEstimatedTimeSeconds={5 * 60}
-      activityComplete={hasCompleted}
+      activityComplete={hasCompletedTraining}
       onSelectionAction={() => navigate('/practice')}
     />
   );
